向量入库

+++ server/sparkai-service/src/main/java/sparkai/service/vo/document/DocumentListVo.java
@@ -50,7 +50,7 @@ public class DocumentListVo implements Serializable {
     /**
      * 分段数
      */
-    private Long paragraphNum;
+    private Integer paragraphNum;
 
     /**
      * 创建时间
